[CHANGE] Some minor eslint refactors. [ADD] isMobile method to pxcontext. [FIX] When request comes from mobile, doesn't use first mode, even if it's enabled. [FIX] Collector url now changes when in first party mode (and not mobile). [ADD] tests for these changes.
