+++ lib/pxcontext.js
@@ -23,7 +23,7 @@ class PxContext {
         this.cookieOrigin = "cookie";
 
         const mobileHeader = this.headers[mobileSdkHeader];
-        if (mobileHeader != undefined) {
+        if (mobileHeader !== undefined) {
             this.cookieOrigin = "header";
             pxLogger.debug('Mobile SDK token detected');
             this.originalToken = this.headers[mobileSdkOriginalTokenHeader];
@@ -74,20 +74,24 @@ class PxContext {
      * @return {boolean} flag sensitive_route true/false.
      */
     checkSensitiveRoute(sensitiveRoutes, uri) {
-        return sensitiveRoutes.some(sensitiveRoute => uri.startsWith(sensitiveRoute) );
+        return sensitiveRoutes.some(sensitiveRoute => uri.startsWith(sensitiveRoute));
     }
 
     getTokenObject(cookie, delimiter = ":") {
         if (cookie.indexOf(delimiter) > -1) {
             let [version, ...extractedCookie] = cookie.split(delimiter);
             if (version === "3") {
-                return {key:"_px3", value: extractedCookie.join(delimiter)};
+                return {key: "_px3", value: extractedCookie.join(delimiter)};
             }
             if (version === "1") {
-                return {key:"_px", value: extractedCookie.join(delimiter)};
+                return {key: "_px", value: extractedCookie.join(delimiter)};
             }
         }
-        return {key:"_px3", value: cookie};
+        return {key: "_px3", value: cookie};
+    }
+
+    isMobile() {
+        return (this.cookieOrigin === 'header');
     }
 }
