+++ lib/pxenforcer.js
@@ -6,9 +6,9 @@ const PxContext = require('./pxcontext');
 const PxConfig = require('./pxconfig');
 const mu = require('mu2');
 const pxProxy = require('./pxproxy');
-const pxHttpc = require('./pxhttpc');
 
 class PxEnforcer {
+
     constructor(params, client) {
         if (client) {
             this.pxClient = client;
@@ -66,7 +66,7 @@ class PxEnforcer {
             pxConfig.ADDITIONAL_ACTIVITY_HANDLER(pxCtx, pxConfig);
         }
 
-        if (pxConfig.CUSTOM_REQUEST_HANDLER){
+        if (pxConfig.CUSTOM_REQUEST_HANDLER) {
             if (res) {
                 pxConfig.CUSTOM_REQUEST_HANDLER(pxCtx, pxConfig, req, res);
                 if (res.headersSent) {
@@ -90,7 +90,7 @@ class PxEnforcer {
             }
 
             pxLogger.debug(`Enforcing action: ${pxCtx.fullBlockAction} page is served`);
-            this.generateTemplate(pxCtx, pxConfig, function(htmlTemplate) {
+            this.generateTemplate(pxCtx, pxConfig, function (htmlTemplate) {
                 const response = {};
                 if (pxCtx.blockAction === 'r') {
                     response.status = '429';
@@ -100,11 +100,11 @@ class PxEnforcer {
                     response.statusDescription = "Forbidden";
                 }
 
-                if (pxCtx.cookieOrigin == "cookie") {
-                    response.header = {key: 'Content-Type', value:'text/html'};
+                if (pxCtx.cookieOrigin === "cookie") {
+                    response.header = {key: 'Content-Type', value: 'text/html'};
                     response.body = htmlTemplate;
                 } else {
-                    response.header = {key: 'Content-Type', value:'application/json'};
+                    response.header = {key: 'Content-Type', value: 'application/json'};
                     response.body = {
                         action: pxCtx.blockAction,
                         uuid: pxCtx.uuid,
@@ -142,6 +142,7 @@ class PxEnforcer {
     /**
      * pxBlock - block handler, send blocking activity to px and render the block html back to screen
      *
+     * @param pxConfig
      * @param {Object} pxCtx - current request context.
      */
     pxBlock(pxCtx, pxConfig) {
@@ -166,7 +167,7 @@ class PxEnforcer {
      * @param {Object} pxConfig - current Px configs
      * @param {Function} cb - send the generated html value.
      */
-    generateTemplate(pxContext, pxConfig, cb){
+    generateTemplate(pxContext, pxConfig, cb) {
         let template;
         switch (pxContext.blockAction) {
             case 'j':
@@ -190,7 +191,7 @@ class PxEnforcer {
             return this.compileMustache('ratelimit', {}, cb);
         }
 
-        if (pxContext.cookieOrigin == 'header') {
+        if (pxContext.isMobile()) {
             template = `${template}.mobile`;
         }
 
@@ -198,14 +199,17 @@ class PxEnforcer {
         return this.compileMustache('block_template', props, cb);
     }
 
-    getProps(pxContext, pxConfig, template){
+    getProps(pxContext, pxConfig, template) {
         let jsClientSrc = `//${pxConfig.CLIENT_HOST}/${pxConfig.PX_APP_ID}/main.min.js`;
         let captchaSrc = `//${pxConfig.CAPTCHA_HOST}/${template}.js`;
-        if (pxConfig.FIRST_PARTY_ENABLED) {
+        let hostUrl = pxContext.collectorUrl;
+
+        if (pxConfig.FIRST_PARTY_ENABLED && !pxContext.isMobile()) {
             const prefix = pxConfig.PX_APP_ID.substring(2).toLowerCase();
             jsClientSrc = `/${prefix}${pxConfig.FIRST_PARTY_VENDOR_PATH}`;
             captchaSrc = `/${prefix}${pxConfig.FIRST_PARTY_CAPTCHA_PATH}/${template}.js`;
-        };
+            hostUrl = `/${prefix}${pxConfig.FIRST_PARTY_XHR_PATH}`;
+        }
 
         return {
             refId: pxContext.uuid,
@@ -216,7 +220,7 @@ class PxEnforcer {
             cssRef: pxConfig.CSS_REF,
             jsRef: pxConfig.JS_REF,
             logoVisibility: pxConfig.CUSTOM_LOGO ? 'visible' : 'hidden',
-            hostUrl: pxContext.collectorUrl,
+            hostUrl: hostUrl,
             jsClientSrc: jsClientSrc,
             firstPartyEnabled: pxConfig.FIRST_PARTY_ENABLED,
             captchaType: pxConfig.CAPTCHA_PROVIDER,
@@ -241,7 +245,6 @@ class PxEnforcer {
         });
     }
 }
-
 
 
 module.exports = PxEnforcer;
